+++ Simulator/Core/Serializer.cpp
@@ -12,10 +12,11 @@
 #include "Serializer.h"
 #include "ConnGrowth.h"
 #include "Connections.h"
+#include "GPUModel.h"
 #include "Simulator.h"
 #include <fstream>
 
-// Displays <Graphitti></Graphitti> as top most element instead of the default <Cereal></Cereal>
+// Displays Graphitti as top most element instead of the default Cereal
 // CEREAL_XML_STRING_VALUE should be placed before defining cereal archives library
 #define CEREAL_XML_STRING_VALUE "Graphitti"
 #include <cereal/archives/binary.hpp>
